Flip HIGH/LOW for SignalConsumer to work with relays
This patch addresses an issue with the specific relay used for
switching the different lamps on/off. The default LOW state for the
relay closes the circuit, which in combination with the 'off' state
for the relay, causes everything on while the arduino boots and
overloads the power supply. Now, the HIGH/LOW is flipped to handle
this (which also requires the hardware config be updated).