Turn off power before switching colors, adjust timing, and move file
This patch adds a delay between switching the aspect/color of the
signal to avoid shorting the power supply (even momentarily). This is
the result of using relays to flip the polarity of the circuit. The
source file has also been moved to more appropriately named directory
(and make the arduino IDE happy).